+++ assets/src/api.ts
@@ -326,6 +326,20 @@ export const fetchClosedConversations = async (token = getAccessToken()) => {
     .then((res) => res.body.data);
 };
 
+export const fetchConversation = async (
+  id: string,
+  token = getAccessToken()
+): Promise<Conversation> => {
+  if (!token) {
+    throw new Error('Invalid token!');
+  }
+
+  return request
+    .get(`/api/conversations/${id}`)
+    .set('Authorization', token)
+    .then((res) => res.body.data);
+};
+
 export const updateConversation = async (
   conversationId: string,
   updates: any,
